+++ industry/nxmodbus/nxmb_internal.h
@@ -205,7 +205,7 @@ static inline uint64_t nxmb_util_clock_ms(void)
   struct timespec ts;
 
   clock_gettime(CLOCK_MONOTONIC, &ts);
-  return (uint64_t)ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+  return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
 }
 
 /****************************************************************************
